LayoutTests/editing/selection/getRangeAt-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
This tests that Selection::getRangeAt(int) returns a valid Range object.
| "\n"
| <div>
|   id="test"
|   style="border:1px solid black; padding:1em;"
|   "[range #0 start]<#selection-anchor>hello<#selection-focus>[range #0 end]"
| "\n\n\n"
